articular-examples: added architecture component-diagram for TestArticularJme
