+++ articular-examples/src/main/java/articular/example/labs/techdemos/jme/testjaime/components/package-info.java
@@ -29,4 +29,4 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package articular.example.labs.techdemos.jme.systems;
+package articular.example.labs.techdemos.jme.testjaime.components;
