+++ articular-examples/src/main/java/articular/example/labs/techdemos/jme/testjaime/package-info.java
@@ -29,24 +29,4 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package articular.example.labs.techdemos.jme.components;
-
-import articular.core.component.Component;
-import com.jme3.input.controls.KeyTrigger;
-
-public final class InputComponent implements Component {
-
-    private final String mapping;
-    private final KeyTrigger key;
-
-    public InputComponent(String mapping, KeyTrigger key) {
-        this.mapping = mapping;
-        this.key = key;
-    }
-
-    @Override
-    public Id getId() {
-        return new Component.Id((hashCode() >>> 16) ^
-                GameComponents.INPUT_COMPONENTS.getEntity().getId().intValue());
-    }
-}
+package articular.example.labs.techdemos.jme.testjaime;
